+++ src/board/edges.rs
@@ -11,15 +11,15 @@ pub struct Edges {
 }
 
 pub fn generate_edges(buffer: EdgeBuffer, rows: Rows, cols: Cols) -> Edges {
-    let top = !(BitBoard::max() << (buffer * cols));
-    let bottom = BitBoard::max() << ((rows - buffer) * cols);
+    let top = !(BitBoard::max() << (buffer * cols) as u32);
+    let bottom = BitBoard::max() << ((rows - buffer) * cols) as u32;
 
-    let mut left = BitBoard::max() & &(!(BitBoard::max() << buffer));
+    let mut left = BitBoard::max() & &(!(BitBoard::max() << (buffer as u32)));
     for _ in 1..rows {
-        left |= &(left << cols);
+        left |= &(left << (cols as u32));
     }
 
-    let right = left << (cols - buffer);
+    let right = left << (cols - buffer) as u32;
 
     let edges = top | &bottom | &left | &right;
